Added support for flags
So far, you use `wego 1 Paris` to fetch weather data for Paris for 1
day. With this commit, you use `wego -days 1 -city Paris` to fetch the
same information. Benefit of this approach is that (1) it prints help
message when you do `wego -h` (2) it is clear what data you're seeking.