Minor changes in ascii art (#155)
* Make sun in "partly cloudy" less irregular

* Make "sunny" more intense

I have noticed that when there's a lot of "sunny" and "partly cloudy" in
the output, it's not apparent at first sight which is which. This should
make the "sunny" icon more intensely yellow, and therefore more
discernible.

* Make dark grey and dark blue a bit ligther

Previous colors were not readable against some background colors,
e.g. #0a3947. New colors are more readable but should be still
sufficiently darker than the ones used for light clouds (250) and light
rain (111).